week.1.1/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pypdf==5.0.1
rank-bm25==0.2.2
